Refactor code and start on music integration
* All non-repo files in the root folder except `index.html` have been
moved to `src/`. This is so that users will know what file to open when
they open the folder.
* Fixed a number of bugs involving communication between the parent and
child windows, and regarding playing and pausing not working on various
windows.
* Created `.gitignore`. Files ignored will include `music/*` - this is
to avoid any copyright issues regarding distribution of music files.
* Start work on music integration. If a music file is not detected, the
Music selection will not appear on the settings menu.